requirements.txt
Pillow
tweepy
python-dotenv
requests
